fix focus event
